requirements.txt
#Fabric
#ecdsa
django
django-debug-toolbar
django-audit-log
jsonfield
django-auth-ldap
djangorestframework
pyyaml
markdown
django-filter
